Fix build warnings and space key not working in TextFields
Replace Form/.formStyle(.grouped) with plain VStack/Grid layout in
AddSourceView to fix space characters not rendering until the next
keystroke — a macOS 26 SwiftUI Form cell-rendering issue.

Also fix NowPlayingManager to properly remove MPRemoteCommandCenter
targets and disable commands on clearNowPlaying(), preventing stale
media key interception after playback ends.

Clean build warnings: silence unused parameter, add explicit
int64->double casts, remove unused metal_video_renderer.mm from build.

+++ core/src/demuxer/ff_demuxer.cpp
@@ -38,7 +38,7 @@ Error FFDemuxer::open(const std::string& path) {
     populate_media_info();
     PY_LOG_INFO(TAG, "Opened: %s (%s), duration=%.2fs, %zu tracks",
                 path.c_str(), info_.container_format.c_str(),
-                info_.duration_us / 1e6, info_.tracks.size());
+                static_cast<double>(info_.duration_us) / 1e6, info_.tracks.size());
 
     return Error::Ok();
 }
